+++ app/models/hammerstone/refine/filter.rb
@@ -23,6 +23,7 @@ module Hammerstone::Refine
         @blueprint = blueprint
         @relation = initial_query
         @immediately_commit_pending_relationship_subqueries = false
+        @@default_stabilizer = Hammerstone::Refine::Stabilizers::UrlEncodedStabilizer
       end
     end
 
@@ -46,10 +47,15 @@ module Hammerstone::Refine
       @initial_query
     end
 
+    def automatically_stabilize?
+      true
+    end
+
     # If the initial condition has not been set in the filter, validation for relationships will error.
     # The BT implementation builds the filter *then* sets the initial condition
     # This pulls a "smart" default from the filter arel table - Arel::Table provides the table name, can extract the model then call "all"
     # to return and AR::R object which is what `uses_attributes` expects.
+    # TODO: This breaks for namespaced classes.
     def fallback_initial_condition
       # TODO: Figure out a better way to do this.
       table.name.classify.constantize.all