+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 ### 2.4.3 (not yet released)
   * bugfix
     * `Clause doesn't equal` and `clause not between` were not implemented for Date with time conditions. PR #98
+  * Clarified the installation documentation
+  * Included the StoredFilter class in the gem assuming that the client would override it as
+  necessary
 
 ### 2.4.2
   * Update the configuration of stabilizer classes