+++ CHANGELOG.md
@@ -1,4 +1,6 @@
 ### 2.4.3 (not yet released)
+  * features
+    * added generator for creating filters
   * bugfix
     * `Clause doesn't equal` and `clause not between` were not implemented for Date with time conditions. PR #98