+++ CHANGELOG.md
@@ -2,6 +2,9 @@
   * bugfix
     * `Clause doesn't equal` and `clause not between` were not implemented for Date with time conditions. PR #98
     * Add in missing has_error classes as specified by Tamik
+  * Clarified the installation documentation
+  * Included the StoredFilter class in the gem assuming that the client would override it as
+  necessary
 
 ### 2.4.2
   * Update the configuration of stabilizer classes