+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ### 2.4.3 (not yet released)
   * bugfix
     * `Clause doesn't equal` and `clause not between` were not implemented for Date with time conditions. PR #98
+    * Add in missing has_error classes as specified by Tamik
   * Clarified the installation documentation
   * Included the StoredFilter class in the gem assuming that the client would override it as
   necessary