+++ CHANGELOG.md
@@ -1,3 +1,18 @@
+### 2.4.3 (not yet released)
+  * features
+    * added generator for creating filters
+  * bugfix
+    * `Clause doesn't equal` and `clause not between` were not implemented for Date with time conditions. PR #98
+    * Add in missing has_error classes as specified by Tamik
+    * Fix raw_attribute bug for text conditions
+  * Clarified the installation documentation
+  * Included the StoredFilter class in the gem assuming that the client would override it as
+  necessary
+  * Default stored filters to false in filter_builder_dropdown
+  * Add automatically stabilize to the gem because it is always required 
+  * Add the filter pill partial view to the gem (not to docs because CSS is still an issue)
+  * Set default stabilizers to be URL Encoded in filter.rb
+
 ### 2.4.2
   * Update the configuration of stabilizer classes